[REFACTOR] further store vs ui decoupling
